requirements.txt
zeroconf==0.131.0
cryptography==38.0.3
argon2-cffi==23.1.0
